requirements.txt
openai-whisper>=20231117
moviepy>=1.0.3
numpy>=1.24.0
torch>=2.0.0
transformers>=4.30.0
scikit-learn>=1.0.2
nltk>=3.8.1
python-dotenv>=1.0.0
tqdm>=4.65.0
pydub>=0.25.1 